#67 - change folder structure
