+++ components/Admin/AboutDetails/EditAboutDetails.tsx
@@ -2,49 +2,49 @@ import React, { useState } from 'react';
 import { FormProvider, SubmitHandler, useForm } from 'react-hook-form';
 import Modal from 'react-modal';
 import { yupResolver } from '@hookform/resolvers/yup';
-import { BsPlusLg } from 'react-icons/bs';
-import customStyles from '../../styles/modalStyles';
-import SubmitButton from './SubmitButton';
-import Textarea from './Textarea';
-import Input from './Input';
-import FormWrapper from './FormWrapper';
-import { useAddAboutDetail } from '../../lib/operations/about';
-import { AboutDetail } from '../../lib/types';
-import { aboutDetailSchema } from '../../lib/schemas';
-import ImageUploader from './ImageUploader';
+import customStyles from '../../../styles/modalStyles';
+import SubmitButton from '../Form/SubmitButton';
+import Textarea from '../Form/Textarea';
+import Input from '../Form/Input';
+import FormWrapper from '../Form/FormWrapper';
+import { useUpdateAboutDetail } from '../../../lib/operations/about';
+import { AboutDetail } from '../../../lib/types';
+import { aboutDetailSchema } from '../../../lib/schemas';
+import ImageUploader from '../Form/ImageUploader';
 
-const AddAboutDetails = () => {
+interface EditAboutDetailsProps {
+  aboutDetail: AboutDetail;
+}
+
+const EditAboutDetails: React.FC<EditAboutDetailsProps> = ({ aboutDetail }) => {
   const [isModalOpen, setIsModalOpen] = useState(false);
   const methods = useForm<AboutDetail>({
     resolver: yupResolver(aboutDetailSchema),
     mode: 'onChange',
   });
-  const { handleSubmit, reset } = methods;
+  const { handleSubmit } = methods;
 
-  const { mutate, isLoading, isError } = useAddAboutDetail();
+  const { mutate, isLoading, isError } = useUpdateAboutDetail(aboutDetail.id);
 
   const onSubmit: SubmitHandler<AboutDetail> = async (data) => {
     const { image, title, description } = data;
     const form = new FormData();
-
     form.append('data', JSON.stringify({ title, description }));
     if (image) form.append('image', image);
 
     mutate(form);
     if (!isError) {
       setIsModalOpen(false);
-      reset();
     }
   };
 
   return (
     <div>
       <button
-        className="text-white bg-black px-5 py-2 rounded-lg"
+        className="text-white bg-green-500 px-5 py-1 rounded-lg"
         onClick={() => setIsModalOpen(true)}
       >
-        <BsPlusLg className="inline mb-1 mr-3" />
-        Add about detail
+        Edit
       </button>
       <Modal
         isOpen={isModalOpen}
@@ -54,17 +54,28 @@ const AddAboutDetails = () => {
         <FormProvider {...methods}>
           <FormWrapper onSubmit={handleSubmit(onSubmit)} multipart>
             <h2 className="text-center text-2xl pb-8 -mt-5">
-              Add about detail
+              Edit about detail
             </h2>
             <div className="mb-5">
-              <ImageUploader id="about-image" label="Image" width={200} />
+              <ImageUploader
+                id="about-image"
+                label="Image"
+                width={200}
+                defaultImage={aboutDetail.image}
+                loader
+              />
             </div>
-            <Input />
+            <Input defaultValue={aboutDetail.title} />
             <div className="mt-5">
-              <Textarea id="description" title="Description" rows={10} />
+              <Textarea
+                id="description"
+                title="Description"
+                rows={10}
+                defaultValue={aboutDetail.description}
+              />
             </div>
             <div className="flex justify-center mt-5">
-              <SubmitButton name="Save" isLoading={isLoading} />
+              <SubmitButton name="Update" isLoading={isLoading} />
             </div>
           </FormWrapper>
         </FormProvider>
@@ -73,4 +84,4 @@ const AddAboutDetails = () => {
   );
 };
 
-export default AddAboutDetails;
+export default EditAboutDetails;
