+++ components/Admin/Form/StatusToggler.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { Switch } from '@headlessui/react';
 import { useFormContext } from 'react-hook-form';
-import camelize from '../../utils/camelize';
+import camelize from '../../../utils/camelize';
 
 interface StatusTogglerProps {
   id: string;
