+++ components/Admin/index.ts
@@ -1,18 +1,18 @@
-export { default as AddButton } from './AddButton';
-export { default as BackButton } from './BackButton';
-export { default as EditableList } from './EditableList';
-export { default as Entries } from './Entries';
-export { default as FormWrapper } from './FormWrapper';
-export { default as GuestStatusToggler } from './StatusToggler';
-export { default as Header } from './Header';
-export { default as HousekeepingComments } from './HousekeepingComments';
-export { default as HousekeepingStatus } from './HousekeepingStatus';
-export { default as ImagesUploader } from './ImagesUploader';
-export { default as ImageUploader } from './ImageUploader';
-export { default as Input } from './Input';
-export { default as PriorityStatus } from './PriorityStatus';
-export { default as SelectInput } from './SelectInput';
-export { default as SubmitButton } from './SubmitButton';
-export { default as Table } from './Table';
-export { default as Textarea } from './Textarea';
-export { default as StatusToggler } from './StatusToggler';
+export { default as AddButton } from './Table/AddButton';
+export { default as BackButton } from './Table/BackButton';
+export { default as EditableList } from './Form/EditableList';
+export { default as Entries } from './Table/Entries';
+export { default as FormWrapper } from './Form/FormWrapper';
+export { default as GuestStatusToggler } from './Form/StatusToggler';
+export { default as Header } from './Table/Header';
+export { default as HousekeepingComments } from './Housekeeping/HousekeepingComments';
+export { default as HousekeepingStatus } from './Housekeeping/HousekeepingStatus';
+export { default as ImagesUploader } from './Form/ImagesUploader';
+export { default as ImageUploader } from './Form/ImageUploader';
+export { default as Input } from './Form/Input';
+export { default as PriorityStatus } from './Housekeeping/PriorityStatus';
+export { default as SelectInput } from './Form/SelectInput';
+export { default as SubmitButton } from './Form/SubmitButton';
+export { default as Table } from './Table/Table';
+export { default as Textarea } from './Form/Textarea';
+export { default as StatusToggler } from './Form/StatusToggler';
