+++ pages/admin/guests/index.tsx
@@ -2,10 +2,10 @@ import React from 'react';
 import { dehydrate, QueryClient } from 'react-query';
 import moment from 'moment';
 import { Entries } from '../../../components/Admin';
-import AddButton from '../../../components/Admin/AddButton';
-import DeleteButton from '../../../components/Admin/DeleteButton';
-import EditButton from '../../../components/Admin/EditButton';
-import Header from '../../../components/Admin/Header';
+import AddButton from '../../../components/Admin/Table/AddButton';
+import DeleteButton from '../../../components/Admin/Table/DeleteButton';
+import EditButton from '../../../components/Admin/Table/EditButton';
+import Header from '../../../components/Admin/Table/Header';
 import Seo from '../../../components/Seo';
 import { getGuests } from '../../../lib/api/guests';
 import { useDeleteGuest, useGetGuests } from '../../../lib/operations/guests';
